% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_Arima.R
\name{Cripto_Arima}
\alias{Cripto_Arima}
\title{Criptocurrency Arima (Modelo Autorregresivo Integrado de Media Movil)}
\usage{
Cripto_Arima(df, temp, n_predict)
}
\arguments{
\item{df}{A data.frame with time series}

\item{temp}{Interval of time series (DAILY, WEEKLY, MONTHLY)}

\item{n_predict}{The intervals to predict value}
}
\value{
One plot of Time Series prediction using ARIMA
}
\description{
Criptocurrency Arima (Modelo Autorregresivo Integrado de Media Movil)
}
\examples{

ETH <- Cripto_time_series(cripto = "ETH", temp = "MONTHLY")
Cripto_Arima(ETH, "MONTHLY", 6)

}
\author{
Erick Cuevas Fernández
}
