% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_exchange.R
\name{Cripto_exchange}
\alias{Cripto_exchange}
\title{Criptocurrency Exchange Rate}
\usage{
Cripto_exchange(cripto = "LTC", exchange = "MXN")
}
\arguments{
\item{cripto}{Specify the symbol of the cryptocurrency}

\item{exchange}{Specifiy the symbol of the market}
}
\value{
This function return  the realtime exchange rate for any pair of digital currency.
}
\description{
Criptocurrency Exchange Rate
}
\examples{

Cripto_exchange("BTC", "MXN")

}
\author{
Erick Cuevas Fernández
}
