% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_Holt_Winters.R
\name{Cripto_Holt_Winters}
\alias{Cripto_Holt_Winters}
\title{Criptocurrency Holt Winters}
\usage{
Cripto_Holt_Winters(df, temp, n_predict)
}
\arguments{
\item{df}{A data.frame with time series}

\item{temp}{Interval of time series (DAILY, WEEKLY, MONTHLY)}

\item{n_predict}{The intervals to predict value}
}
\value{
Two plots of Time Series prediction using HoltWinters
}
\description{
Criptocurrency Holt Winters
}
\examples{

ETH <- Cripto_time_series(cripto = "ETH", temp = "MONTHLY")
Cripto_Holt_Winters(ETH, "MONTHLY", 6)

}
\author{
Erick Cuevas Fernández
}
