% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_plot_predict.R
\name{Cripto_plot_predict}
\alias{Cripto_plot_predict}
\title{Crypto Plot of Predictions}
\usage{
Cripto_plot_predict(df, temp = "DAILY", n_predicted = 10,
  crypto = "BTC")
}
\arguments{
\item{df}{A list with the predictions}

\item{temp}{Units of time predicted}

\item{n_predicted}{Numer of units predicted}

\item{crypto}{A Crypto Symbol}
}
\value{
A list with a data.frame with the predictions, and two plots of the predictions
}
\description{
Crypto Plot of Predictions
}
\examples{

LTC_monthly <- Cripto_time_series(cripto = "LTC", market = "MXN", temp = "MONTHLY")
LTC_hw3months <- Cripto_Holt_Winters(df = LTC_monthly, temp = "MONTHLY", n_predict = 3)
plots_LTC <- Cripto_plot_predict(df = LTC_hw3months, temp = "MONTHLY", n_predicted = 3, crypto = "LTC")

}
\author{
Erick Cuevas Fernandez
}
