% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_seasonal_analysis.R
\name{Cripto_seasonal_analysis}
\alias{Cripto_seasonal_analysis}
\title{Criptocurrency Seasonal Analysis}
\usage{
Cripto_seasonal_analysis(df, temp)
}
\arguments{
\item{df}{A data.frame with time series}

\item{temp}{Interval of time series (DAILY, WEEKLY, MONTHLY)}
}
\value{
Two plots of Time Series decomposition and a List with the results of decompisitions
}
\description{
Criptocurrency Seasonal Analysis
}
\examples{

ETH <- Cripto_time_series(cripto = "ETH", temp = "MONTHLY")
Cripto_seasonal_analysis(ETH, "MONTHLY")

}
\author{
Erick Cuevas Fernández
}
