% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cripto_time_series.R
\name{Cripto_time_series}
\alias{Cripto_time_series}
\title{Criptocurrency Time Series}
\usage{
Cripto_time_series(cripto = "LTC", market = "MXN", temp = "WEEKLY")
}
\arguments{
\item{cripto}{Specify the symbol of the cryptocurrency}

\item{market}{Specifiy the symbol of the market}

\item{temp}{DAILY, WEEKLY or MOTHLY}
}
\value{
Return a data.frame the daily historical time series for a digital currency (e.g., BTC) traded on a specific market (e.g., CNY/Chinese Yuan), refreshed daily at midnight (UTC). Prices and volumes are quoted in both the market-specific currency and USD.
}
\description{
Criptocurrency Time Series
}
\examples{
Cripto_time_series(cripto = "ETH", temp = "MONTHLY")


}
\author{
Erick Cuevas Fernández
}
